% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkRvsCpp.R
\name{BenchmarkRvsCpp}
\alias{BenchmarkRvsCpp}
\title{A log-likelihood calculation time comparison for different numbers of traits and option-sets}
\usage{
BenchmarkRvsCpp(ks = c(1, 2, 4, 8), optionSets = NULL,
  verbose = FALSE)
}
\arguments{
\item{ks}{a vector of positive integers, denoting different numbers of traits. 
Default: \code{c(1, 2, 4, 8)}.}

\item{optionSets}{a named list of lists of PCM-options. If NULL (the default) the 
option set is set to \code{
list(
`serial / 1D-multiv.` = list(PCMBase.Lmr.mode = 11, PCMBase.Threshold.EV = 1e-9, PCMBase.Threshold.SV = 1e-9, PCMBase.Use1DClasses = FALSE),
`parallel / 1D-multiv.` = list(PCMBase.Lmr.mode = 21, PCMBase.Threshold.EV = 1e-9, PCMBase.Threshold.SV = 1e-9, PCMBase.Use1DClasses = FALSE),
`serial / 1D-univar.` = list(PCMBase.Lmr.mode = 11, PCMBase.Threshold.EV = 1e-9, PCMBase.Threshold.SV = 1e-9, PCMBase.Use1DClasses = TRUE),
`parallel / 1D-univar.` = list(PCMBase.Lmr.mode = 21, PCMBase.Threshold.EV = 1e-9, PCMBase.Threshold.SV = 1e-9, PCMBase.Use1DClasses = TRUE)
)} for k = 1 and to \code{
list(
`serial / 1D-multiv.` = list(PCMBase.Lmr.mode = 11, PCMBase.Threshold.EV = 1e-9, PCMBase.Threshold.SV = 1e-9, PCMBase.Use1DClasses = FALSE),
`parallel / 1D-multiv.` = list(PCMBase.Lmr.mode = 21, PCMBase.Threshold.EV = 1e-9, PCMBase.Threshold.SV = 1e-9, PCMBase.Use1DClasses = FALSE))} for k > 1.}

\item{verbose}{logical indicating if log-messages should be printed to the console during the benchmark. Default FALSE.}
}
\value{
a data.table for results similar to the data.table returned from \code{\link{MiniBenchmarkRvsCpp}} with 
additional columns for k, option-set and the type of model.
}
\description{
A log-likelihood calculation time comparison for different numbers of traits and option-sets
}
