% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{Lmr.BM}
\alias{Lmr.BM}
\title{Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree}
\usage{
Lmr.BM(model, metaI, pruneI)
}
\arguments{
\item{model}{parameters of the OU process. This must be a
named list with the following elements:
Sigma: a R x k x k array, each Sigma[r,,] containing the
matrix Sigma for regime r;
Sigmae: a R x k x k array, each Sigmae[r,,] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{pruneI}{an object of class QuadraticPolynomialOU, which has been created using
newCppObject.OU.}
}
\value{
a named list containing the following elements:
L: a k x k matrix
m: a k vector
r: a number;
}
\description{
Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree
}
