% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU.R
\name{Lmr.OU}
\alias{Lmr.OU}
\title{Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree}
\usage{
Lmr.OU(model, metaI, pruneI)
}
\arguments{
\item{model}{parameters of the OU process. This must be a
named list with the following elements:
Alpha: a R x k x k array, where R is the number of regimes of the
OU process, k is the number of variables (traits), each Alpha[r,,]
containing the matrix Alpha for regime r;
Theta: a R x k matrix, row Theta[r, ] containing the long-term
mean Theta for regime r;
Sigma: a R x k x k array, each Sigma[r,,] containing the
matrix Sigma for regime r;
Sigmae: a R x k x k array, each Sigmae[r,,] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{pruneI}{an object of class QuadraticPolynomialOU, which has been created using
newCppObject.OU.}
}
\value{
a named list containing the following elements:
L: a k x k matrix
m: a k vector
r: a number;
}
\description{
Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree
}
