% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOU.R
\name{Lmr.Rcpp_QuadraticPolynomialJOU}
\alias{Lmr.Rcpp_QuadraticPolynomialJOU}
\title{Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree}
\usage{
Lmr.Rcpp_QuadraticPolynomialJOU(X, tree, model, metaI = validateModel(tree,
  model), pruneI = newCppObject(X, tree, model, metaI),
  pc = presentCoordinates(X, tree), root.only = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{parameters of the JOU process. This must be a
named list with the following elements:
Alpha: a k x k x R array, where R is the number of regimes of the
JOU process, k is the number of variables (traits), each Alpha[,,r]
containing the matrix Alpha for regime r;
Theta: a k x R matrix, row Theta[, r] containing the long-term
mean Theta for regime r;
Sigma: a k x k x R array, each Sigma[,,r] containing the
matrix Sigma for regime r;
Sigmae: a k x k x R array, each Sigmae[,,r] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{pruneI}{an object of class QuadraticPolynomialJOU, which has been created using
newCppObject.JOU.}
}
\value{
a named list containing the following elements:
L: a k x k matrix
m: a k vector
r: a number;
}
\description{
Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree
}
