% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniBenchmarkData}
\alias{miniBenchmarkData}
\title{Data for performing a mini-benchmark}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{X}{trait values}
  \item{tree}{phylogenetic tree (phylo) with set edge.regimes member}
  \item{model}{PCM model}
}}
\usage{
miniBenchmarkData
}
\description{
A dataset containing ten triplets of trees, trait-values and models to
evaluate the likelihood calculation times for R and C++ implementations.
}
\keyword{datasets}
