% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{newCppObject}
\alias{newCppObject}
\title{A generic method for creating C++ backend objects given a model, data and 
a tree.}
\usage{
newCppObject(X, tree, model, metaInfo = validateModel(tree, model), ...)
}
\description{
A generic method for creating C++ backend objects given a model, data and 
a tree.
}
