% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{PCListInt}
\alias{PCListInt}
\title{Converts the logical matrix pc into a list of vectors denoting the (0-based) TRUE-indices in 
each column}
\usage{
PCListInt(pc)
}
\description{
Converts the logical matrix pc into a list of vectors denoting the (0-based) TRUE-indices in 
each column
}
