% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{PCMCppPruningObject}
\alias{PCMCppPruningObject}
\title{A generic method for creating C++ backend objects given a model, data and 
a tree.}
\usage{
PCMCppPruningObject(X, tree, model, metaInfo = PCMValidate(tree, model), ...)
}
\description{
A generic method for creating C++ backend objects given a model, data and 
a tree.
}
