% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{PCMInfoCpp}
\alias{PCMInfoCpp}
\title{A generic method for creating C++ backend objects given a model, data and 
a tree.}
\usage{
PCMInfoCpp(X, tree, model, metaI = PCMInfo(X, tree, model, verbose, preorder =
  PCMTreePreorderCpp(tree)), verbose = FALSE, ...)
}
\description{
A generic method for creating C++ backend objects given a model, data and 
a tree.
}
