% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSpeedOU.R
\name{PCMLmr.Rcpp_QuadraticPolynomialTwoSpeedOU}
\alias{PCMLmr.Rcpp_QuadraticPolynomialTwoSpeedOU}
\title{Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree}
\usage{
PCMLmr.Rcpp_QuadraticPolynomialTwoSpeedOU(X, tree, model,
  metaI = PCMValidate(tree, model), pruneI = PCMCppPruningObject(X, tree,
  model, metaI), pc = PCMPresentCoordinates(X, tree), root.only = FALSE,
  verbose = FALSE)
}
\arguments{
\item{model}{parameters of the 2SpOU process. This must be a
named list with the following elements:

H1: a k x k x R array, where R is the number of regimes of the
OU process, k is the number of variables (traits), each H1[,,r]
containing the matrix H1 for regime r;

H2: a k x k x R array, where R is the number of regimes of the
OU process, k is the number of variables (traits), each H2[,,r]
containing the matrix H2 for regime r;

Theta: a k x R matrix, row Theta[, r] containing the long-term
mean Theta for regime r;

Sigma: a k x k x R array, each Sigma[,,r] containing the
matrix Sigma for regime r;

Sigmae: a k x k x R array, each Sigmae[,,r] representing a diagonal matrix
with elements on the diagona corresponding to the environmental variances for
the k traits in regime r}

\item{pruneI}{an object of class QuadraticPolynomialOU, which has been created using
PCMCppPruningObject.TwoSpeedOU.}
}
\value{
a named list containing the following elements:
L: a k x k matrix
m: a k vector
r: a number;
}
\description{
Calculate the coefficients L, m, r of the general
form (eq. 2) for the root node of a tree
}
