% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiniBenchmarkRvsCpp.R
\name{MiniBenchmarkRvsCpp}
\alias{MiniBenchmarkRvsCpp}
\title{Evaluate the likelihood calculation times for example trees and data}
\usage{
MiniBenchmarkRvsCpp(data = PCMBaseCpp::miniBenchmarkData,
  nRepsCpp = 10L, listOptions = list(PCMBase.Lmr.mode = 11,
  PCMBase.Threshold.EV = 1e-07))
}
\arguments{
\item{data}{a `data.frame` with at least the following columns: 
\itemize{
\item{tree: }{a list column of phylo objects with an edge.regime member set.}
\item{X: }{a list column of k x N numerical matrices.}
\item{model: }{a list column of PCM objects.}
}
Defaults: to `miniBenchmarkData`, which is small data-table included
with the PCMBaseCpp package.}

\item{nRepsCpp}{: number of repetitions for the cpp likelihood calculation 
calls: a bigger value increases the precision of time estimation at the 
expense of longer running time for the benchmark. Defaults to 10.}

\item{listOptions}{options to set before measuring the calculation times. 
Defaults to `list(PCMBase.Lmr.mode = 11, PCMBase.Threshold.EV = 1e-7)`. 
`PCMBase.Lmr.mode` corresponds to the parallel traversal mode for the tree 
traversal algorithm (see 
\href{https://venelin.github.io/SPLITT/articles/SPLITTTraversalModes.html}{this page}
for possible values).}
}
\value{
a data.frame.
}
\description{
Evaluate the likelihood calculation times for example trees and data
}
\examples{
\dontrun{
library(PCMBase)
library(PCMBaseCpp)
MiniBenchmarkRvsCpp()

MiniBenchmarkRvsCpp(listOptions = list(PCMBase.Lmr.mode = 21))
}
}
